moved tags related code to services/utils/search/tags.js
